Renamed postgres.connection to db.connection
