+++ src/core/index.ts
@@ -4,4 +4,4 @@ export * from './base.cache';
 export * from './base.modelclass';
 export * from './base.options';
 export * from './base.service';
-export * from './postgres.connection';
+export * from './db.connection';
